# Evaluate triggerables for all repeat instances where needed