--- a/src/main/java/org/javarosa/core/model/TriggerableDag.java
+++ b/src/main/java/org/javarosa/core/model/TriggerableDag.java
@@ -645,10 +645,6 @@ public class TriggerableDag {
         // step, but i think it's supposed to work
     }
 
-    boolean shouldTrustPreviouslyCommittedAnswer() {
-        return false;
-    }
-
     final void publishSummary(String lead, TreeReference ref, Collection<QuickTriggerable> quickTriggerables) {
         accessor.getEventNotifier().publishEvent(new Event(lead + ": " + (ref != null ? ref.toShortString() + ": " : "") + quickTriggerables.size() + " triggerables were fired."));
     }