@ggalmazor 2020-06-21
Impressive work both on the implementation side and the effort for having every corner case caracterized in tests! 👍 👍 👍 👍 👍 

We will have a chat in a couple of minutes, but I want to write down some thoughts/comments:

> Does initializeTriggerables need to consider triggerables that affect all instances? addingRepeatInstance_withInnerCalculateDependentOnOuterSum_updatesInnerSumForAllInstances() gets at something related.

I think it does, but only if we want to cover that corner case. It doesn't look like something you would find very often, so it could be OK to document it and ignore it for now.

> One case I know will be slower is when there's a position(..) expression and a new repeat instance is added

I suggest we add a benchmark around that to keep it under control.

> I think that the worst case would be that firedAnchors covers some case we haven't identified. 

We've already established that the original implementation can't be explained. We even guess that it's an accidental implementation.

I think that replacing the implementation is something that we can't really avoid. Now the question would be: how hard is going to be to add a new scenario we don't want know about?

My opinion is that now we would be much better suited to change the DAG's behavior than in v2.17. This is not only because we've learned about the DAG's internals, but also because the code has a better design and there's a huge amount of tests that will let us work safely and with confidence.

> The big change in behavior compared to v2.17.0 is that adding a repeat instance will trigger recomputation of any calculation inside that repeat which refers to the generic repeat such as count(../../repeat) or position(..). I don't think anyone could have been relying on that bad behavior.

Let's document this and provide form designers with alternative implementations that follow best practices.